reorganize this into a sensible tree
these files will need to be moved or turned into an overlay if they are
to be applied from here, but we will wind up creating a cluster (fleet)
repo during the Weave GitOps installation anyway, since we are lucky and
are installing WGE which has CAPI support as a first-class feature. 🎉😎

Some of these files may not get used at all. This is simply a process of
going from "we use VCluster and we maintain our configuration in Git" to
"we actually use Cluster API and GitOps together" a completely different
ballgame and strategy to scale

Signed-off-by: Kingdon Barrett <kingdon@weave.works>